Feat(item image): 상품 이미지 수정 api

+++ backend/yam/src/main/java/com/ssafy/yam/domain/item/service/ItemCRUDService.java
@@ -3,6 +3,7 @@ package com.ssafy.yam.domain.item.service;
 import com.ssafy.yam.domain.image.entity.Image;
 import com.ssafy.yam.domain.image.repository.ImageRepository;
 import com.ssafy.yam.domain.item.dto.request.ItemCreateRequest;
+import com.ssafy.yam.domain.item.dto.request.ItemImageRequest;
 import com.ssafy.yam.domain.item.dto.request.ItemUpdateRequest;
 import com.ssafy.yam.domain.item.entity.Item;
 import com.ssafy.yam.domain.item.repository.ItemRepository;
@@ -47,28 +48,9 @@ public class ItemCRUDService {
         item.setItemModifiedTime(LocalDateTime.now());
         item.setItemIsActive("Y");
 
-//        item.setUser(userService.getUser(itemCreateRequest.getUid()));
         itemRepository.save(item);
 
-        //if (itemImages.size() != 0) {
-            for (MultipartFile itemImage : itemImages) {
-                if(!itemImage.isEmpty()) {
-                    String imageUrl = null;
-                    String userSet = itemImage.getOriginalFilename() + "(" + LocalDate.now().toString() + ")";
-                    try {
-                        imageUrl = s3UploadUtils.upload(itemImage, "item", userSet);
-//                    System.out.println(itemImage.getOriginalFilename() + " : profile image upload s3 success");
-                    } catch (IOException e) {
-//                    System.out.println(itemImage.getOriginalFilename() + " : profile image upload s3 fail");
-                        e.printStackTrace();
-                    }
-                    Image image = new Image();
-                    image.setImageUrl(imageUrl);
-                    image.setItem(item);
-                    imageRepository.save(image);
-                }
-            }
-        //}
+        uploadImage(itemImages, item);
     }
 
     private Item getItem(int itemId) {
@@ -76,26 +58,57 @@ public class ItemCRUDService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 제품이 존재하지 않습니다."));
     }
 
-    public void deleteItem(int itemId){
-//        Item item = getItem(itemId);
-//        if (!item.getItem.getUid().equals(uid)) {
-//            throw new IllegalArgumentException("물품을 삭제할 권한이 없습니다.");
-//        }
+    public void deleteItem(String token, int itemId){
+        Item item = getItem(itemId);
+        String tokenEmail = TokenUtils.getUserEmailFromToken(token);
+        if (!item.getSeller().getUserEmail().equals(tokenEmail)) {
+            throw new IllegalArgumentException("물품을 삭제할 권한이 없습니다.");
+        }
         itemRepository.deleteById(itemId);
     }
 
-    public void updateItem(ItemUpdateRequest itemUpdateRequest){
-        Item item = modelMapper.map(itemUpdateRequest, Item.class);
-//        item.updateItem(itemUpdateRequest);
+    public void updateItem(String token, ItemUpdateRequest itemUpdateRequest){
+        Item item = getItem(itemUpdateRequest.getItemId());
+        String tokenEmail = TokenUtils.getUserEmailFromToken(token);
+        if (!item.getSeller().getUserEmail().equals(tokenEmail)) {
+            throw new IllegalArgumentException("물품을 수정할 권한이 없습니다.");
+        }
 
-        Optional<Item> updateItem = itemRepository.findById(item.getItemId());
-        updateItem.ifPresent(selectItem ->{
-            selectItem.setItemName(item.getItemName());
-            selectItem.setItemContent(item.getItemContent());
-            selectItem.setItemCategory(item.getItemCategory());
-            selectItem.setItemPrice(item.getItemPrice());
-            selectItem.setItemModifiedTime(LocalDateTime.now());
-            itemRepository.save(selectItem);
-        });
+        item.setItemName(itemUpdateRequest.getItemName());
+        item.setItemContent(itemUpdateRequest.getItemContent());
+        item.setItemCategory(itemUpdateRequest.getItemCategory());
+        item.setItemPrice(itemUpdateRequest.getItemPrice());
+        item.setItemModifiedTime(LocalDateTime.now());
+        itemRepository.save(item);
+    }
+
+    public void addItemImage(String token, int itemId, List<MultipartFile> itemImages){
+        Item item = getItem(itemId);
+        String tokenEmail = TokenUtils.getUserEmailFromToken(token);
+        if (!item.getSeller().getUserEmail().equals(tokenEmail)) {
+            throw new IllegalArgumentException("물품을 수정할 권한이 없습니다.");
+        }
+
+        uploadImage(itemImages, item);
+    }
+
+    private void uploadImage(List<MultipartFile> itemImages, Item item) {
+        for (MultipartFile itemImage : itemImages) {
+            if(!itemImage.isEmpty()) {
+                String imageUrl = null;
+                String userSet = itemImage.getOriginalFilename() + "(" + LocalDate.now().toString() + ")";
+                try {
+                    imageUrl = s3UploadUtils.upload(itemImage, "item", userSet);
+//                    System.out.println(itemImage.getOriginalFilename() + " : profile image upload s3 success");
+                } catch (IOException e) {
+//                    System.out.println(itemImage.getOriginalFilename() + " : profile image upload s3 fail");
+                    e.printStackTrace();
+                }
+                Image image = new Image();
+                image.setImageUrl(imageUrl);
+                image.setItem(item);
+                imageRepository.save(image);
+            }
+        }
     }
 }
